backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic[email]==2.5.0
firebase-admin==6.2.0
google-cloud-firestore==2.13.1
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
